+++ src/common/sockdiag.c
@@ -19,6 +19,7 @@
 
 #include <config.h>
 #include <errno.h>
+#include <inttypes.h>
 #include <stdio.h>
 #include <string.h>
 #include <unistd.h>
@@ -192,7 +193,7 @@ static int receive_responses(int fd, process_response process, void *context)
 
 				if (h->nlmsg_len < NLMSG_LENGTH(sizeof(*err))) {
 					syslog(LOG_ERR,
-					       "nlmsg_type NLMSG_ERROR has short nlmsg_len %d",
+					       "nlmsg_type NLMSG_ERROR has short nlmsg_len %" PRIu32,
 					       h->nlmsg_len);
 				} else {
 					syslog(LOG_ERR, "NLM query failed %s",
@@ -203,8 +204,8 @@ static int receive_responses(int fd, process_response process, void *context)
 			}
 
 			if (h->nlmsg_type != SOCK_DIAG_BY_FAMILY) {
-				syslog(LOG_ERR, "unexpected nlmsg_type %u\n",
-				       (unsigned)h->nlmsg_type);
+				syslog(LOG_ERR, "unexpected nlmsg_type %" PRIu16 "\n",
+				       h->nlmsg_type);
 				return -1;
 			}
 
@@ -212,13 +213,13 @@ static int receive_responses(int fd, process_response process, void *context)
 
 			if (h->nlmsg_len < NLMSG_LENGTH(sizeof(*diag))) {
 				syslog(LOG_ERR,
-				       "nlmsg_type SOCK_DIAG_BY_FAMILY has short nlmsg_len %d",
+				       "nlmsg_type SOCK_DIAG_BY_FAMILY has short nlmsg_len %" PRIu32,
 				       h->nlmsg_len);
 				return -1;
 			}
 
 			if (diag->udiag_family != AF_UNIX) {
-				syslog(LOG_ERR, "unexpected family %u\n",
+				syslog(LOG_ERR, "unexpected family %" PRIu8 "\n",
 				       diag->udiag_family);
 				return -1;
 			}
